+++ src/main/java/ru/top/project/model/Review.java
@@ -1,8 +1,10 @@
 package ru.top.project.model;
+import java.math.BigInteger;
 import java.util.Objects;
 
+
 public class Review {
-    private int id;
+    private BigInteger id;
     Course course ;
     private String comment;
     private String mark;
@@ -33,11 +35,11 @@ public class Review {
         this.comment = comment;
     }
 
-    public int getId() {
+    public BigInteger getId() {
         return id;
     }
     public void setId(int id) {
-        this.id = id;
+        this.id = BigInteger.valueOf(id);
     }
 
     public Course getCourseById() {
